Batch Actions experiment
The idea is to rely on partials that users declare rather than a DSL so full customization is possible. We have to figure out the way of pulling the form inputs back into the form in the paginated collection element. Likely we'll use the same approach but when opening the modal we attach an event listener on any form and detach on close.